Decompile getVar800a236c

+++ src/include/gvars/gvars.h
@@ -705,7 +705,7 @@ extern u16 *g_PadOffsets;
 extern u16 *g_CoverFlags;
 extern s32 *g_CoverRooms;
 extern u32 var800a2368;
-extern u32 var800a236c;
+extern u16 var800a236c;
 extern u32 var800a2370;
 extern struct var800a2380 var800a2380[4];
 extern u32 var800a2eb0;
